Use python scripts as configuration files
Switch the config file from ini format to a python script.

This will be required to let the local administrator or the user add
custom fuctions to specify debuging info ti include in the mail.

Known issues: the python configuration file are executed in the main
script context. This makes monkey-patching very easy for somebody who
could access the user's home. Is it a security issue ?