Clarify online help
It was not clear enought thar the bug reporter must provide *exact steps
how the reproduce the bug*, if possible, and that she should try to
reproduce it. Since we generaly can't get back to them, this is however
vital.

This commit refine the points under "Some help" in the user interface.