Make dependency on python3-pysocks more future proof.
This package was removed from Debian (https://bugs.debian.org/810306)
because it was a duplicate of python3-socksipy.

The binary package is currently called python3-socksipy in Stretch,
but then in sid it was renamed to python3-socks.

So, we depend on any one of those three, but make it clear that we want
1.5.0 or later, since python3-socksipy is only 1.02 on Jessie,
while python3-pysocks (that we have tested WhisperBack with so far on
Jessie) is 1.5.0. As a result, WhisperBack should still pull
python3-pysocks on Jessie, while it'll pull python3-socksipy on Stretch,
and soon python3-socks.